Be more sophisticated about handling SIGINT.
In particular, create a signal handler for SIGINT and note when an
interrupt was received.  Then continue by raising KeyboardInterrupt.
This is useful to handle some code that would fail when presented with
an interrupt but is still forced to executed because it's in an
except/finally block.  Now it can check if an interrupt was received
and fail either gracefully or loudly.  (The reason for this patch was
when a word with an out: parameter was interrupted before its value
could be set.  The code to push the out: value onto the stack ran in
a finally block and complained there was no value --- but it makes
more sense to ignore the whole issue because we're interrupting the
routine anyway.)